stat-filesize needs to accept long long

+++ libstyx/styx.c
@@ -3,11 +3,12 @@
 #include <fcntl.h>
 #include <sys/sendfile.h>
 #include <sys/stat.h>
+#include <sys/types.h>
 #include <openssl/md5.h>
 
 
 
-int styx_stat_filesize(const char* afilename){
+off_t styx_stat_filesize(const char* afilename){
 	struct stat st;
 	stat(afilename, &st);
 	return st.st_size;
